Changes based on feedback
